v3.10 - February 17th, 2018
v3.10 - February 17th, 2018
o Included "Non well-formed number" fix by [url=https://www.simplemachines.org/community/index.php?action=profile;u=112942]GL700Wing[/url], as discussed [url=https://www.simplemachines.org/community/index.php?topic=544111.msg3958439#msg3958439]here[/url].

+++ Subs-AutoRotation.php
@@ -86,8 +86,8 @@ function AutoRotation_Process($filename, $orientation = false)
 		$info['channels'] = !empty($info['channels']) ? $info['channels'] : 4;
 		$memoryNeeded = ceil($width * $height * $info['bits'] * $info['channels'] / 8 * 2);
 		$currentMemory = memory_get_usage() + $memoryNeeded;
-		$memoryLimit = ini_get('memory_limit');
-		$memoryRequested = (int) $memoryLimit + ceil(($currentMemory - $memoryLimit * pow(1024, 2)) / pow(1024, 2));
+		$memoryLimit = (int) ((float) ini_get('memory_limit'));
+		$memoryRequested = $memoryLimit + ceil(($currentMemory - $memoryLimit * pow(1024, 2)) / pow(1024, 2));
 		
 		// Do we have enough memory?  If not, try to allocate enough:
 		if ($currentMemory > (integer) ini_get('memory_limit') * pow(1024, 2))
